requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==65.9.0
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-ai-generativelanguage==0.6.4
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.70.0
googlemaps==4.10.0
grpcio==1.73.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
pillow==10.3.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.0.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
